backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
neo4j==5.28.1
httpx==0.28.1
pydantic-settings==2.8.1
python-dotenv==1.0.1
arxiv==2.1.0
PyPDF2==3.0.1
python-multipart==0.0.9
